pyproject.toml
[tool.poetry]
name = "spotify-weekly-playlist"
version = "0.1.0"
description = ""
authors = ["Vishal Ramesh <vishalramesh50@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
selenium = "^3.141.0"
python-dotenv = "^0.10.5"
requests = "^2.22.0"
pytz = "^2019.3"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
